item imaps corrected, trunk version

+++ .vim/ftplugin/ATP_files/makelatex.py
@@ -4,14 +4,17 @@
 # This file is a part of AutomaticTexPlugin plugin for Vim.
 # It is distributed under General Public Licence v3 or higher.
 
-# To do: compile in temp dir.
+# import signal, os
+# def handler(signum, frame):
 
-import shutil, os.path, re, optparse, subprocess, traceback, psutil, tempfile
+
+import shutil, os.path, re, optparse, subprocess, traceback, psutil
+import tempfile, os, atexit, sys
+
 from optparse import OptionParser
-import os
 from os import getcwd
+from signal import SIGKILL
 from collections import deque
-
 
 usage   = "usage: %prog [options]"
 parser  = OptionParser(usage=usage)
@@ -47,7 +50,7 @@ texfile = options.texfile
 basename = os.path.splitext(os.path.basename(texfile))[0]
 texfile_dir = os.path.dirname(texfile)
 logfile = basename+".log"
-debug_file.write("logfile="+logfile)
+debug_file.write("logfile="+logfile+"\n")
 auxfile = basename+".aux"
 bibfile = basename+".bbl"
 idxfile = basename+".idx"
@@ -56,6 +59,31 @@ tocfile = basename+".toc"
 loffile = basename+".lof"
 lotfile = basename+".lot"
 thmfile = basename+".thm"
+
+if not os.path.exists(os.path.join(texfile_dir, ".tmp")):
+        # This is the main tmp dir (./.tmp) 
+        # it will not be deleted by this script
+        # as another instance might be using it.
+        # it is removed by Vim on exit.
+    os.mkdir(os.path.join(texfile_dir,".tmp"))
+tmpdir  = tempfile.mkdtemp(dir=os.path.join(texfile_dir,".tmp"),prefix="")
+
+# List of pids runing.
+pids    = []
+# Cleanup on exit:
+def cleanup(debug_file, tmpdir, pids):
+    debug_file.close()
+    shutil.rmtree(tmpdir)
+# Will this function be called when scripts get SIGKILL, if yes
+# then this code might be helpful:
+# THIS NEEDS sys MODULE AND SIGKILL from SIGNAL MODULE
+#     for pid in pids:
+#         try:
+#             os.kill(pid,SIGKILL)
+#         except OSError:
+#             # No such process error.
+#             pass
+atexit.register(cleanup, debug_file, tmpdir, pids)
 
 # FILTER:
 def nonempty(str):
@@ -83,19 +111,12 @@ debug_file.write("BIBCMD="+bibcmd+"\n")
 biber=False
 if re.search(bibcmd, '^\s*biber'):
     biber=True
+debug_file.write("BIBER="+str(biber)+"\n")
 bibliographies  = options.bibliographies.split(",")
 bibliographies  = list(filter(nonempty, bibliographies))
 
-tex_options	= options.tex_options
-if re.match('\s*$', tex_options):
-    tex_options_list=[]
-else:
-    if re.search('\s', tex_options):
-        tex_options_list=' '.split(tex_options)
-        tex_options_list=list(filter(nonempty,tex_options_list))
-    else:
-        tex_options_list=[tex_options]
-debug_file.write("TEX_OPTIONS_LIST="+str(tex_options_list)+"\n")
+tex_options     = list(filter(nonempty,re.split('\s*,\s*',options.tex_options)))
+debug_file.write("TEX_OPTIONS_LIST="+str(tex_options)+"\n")
 
 outdir		= options.outdir
 debug_file.write("OUTDIR="+outdir+"\n")
@@ -106,7 +127,7 @@ viewer          = options.viewer
 XpdfServer      = options.xpdf_server
 reload_viewer   = options.reload_viewer
 reload_on_error = options.reload_on_error
-viewer_rawopt   = options.viewer_opt.split(',')
+viewer_rawopt   = re.split('\s*;\s*', options.viewer_opt)
 viewer_it       = list(filter(nonempty,viewer_rawopt))
 viewer_opt      =[]
 for opt in viewer_it:
@@ -116,6 +137,7 @@ if viewer == "xpdf" and XpdfServer != None:
     viewer_opt.extend(["-remote", XpdfServer])
 keep            = options.keep.split(',')
 keep            = list(filter(nonempty, keep))
+debug_file.write("KEEP="+str(keep)+"\n")
 
 def keep_filter_aux(string):
     if string == 'aux':
@@ -164,11 +186,18 @@ def latex_progress_bar(cmd):
 
     child = subprocess.Popen(cmd, stdout=subprocess.PIPE)
     pid   = child.pid
+    pids.append(pid)
 
     vim_remote_expr(servername, "atplib#LatexPID("+str(pid)+")")
     stack = deque([])
     while True:
-        out = child.stdout.read(1).decode()
+        try:
+            out = child.stdout.read(1).decode()
+        except UnicodeDecodeError:
+            debug_file.write("UNICODE DECODE ERROR:\n")
+            debug_file.write(child.stdout.read(1))
+            debug_file.write("\n")
+            out = ""
         if out == '' and child.poll() != None:
             break
         if out != '':
@@ -218,27 +247,35 @@ def reload_xpdf():
 
     if re.search(viewer, '^\s*xpdf\e') and reload_viewer:
         cond=xpdf_server_file_dict().get(XpdfServer, ['_no_file_']) != ['_no_file_'] 
-        if cond and ( reload_on_error or latex_returncode == 0 or bang ):
+        if cond and ( reload_on_error or latex.returncode == 0 or bang ):
             debug_file.write("reloading Xpdf\n")
             run=['xpdf', '-remote', XpdfServer, '-reload']
             devnull=open(os.devnull, "w+")
             subprocess.Popen(run, stdout=devnull, stderr=subprocess.STDOUT)
             devnull.close()
 
-def copy_back_output():
+def copy_back_output(tmpdir):
     os.chdir(tmpdir)
-    shutil.copy(basename+output_ext, texfile_dir)
+    if os.path.exists(file_cp):
+        shutil.copy(basename+output_ext, texfile_dir)
     os.chdir(texfile_dir)
 
-def copy_back():
+def copy_back(tmpdir, latex_returncode):
+
     os.chdir(tmpdir)
-    for ext in list(filter(keep_filter_aux,keep)):
+    if not latex_returncode or not os.path.exists(os.path.join(texfile_dir, auxfile)):
+        ext_list=list(keep)
+    else:
+        ext_list=list(filter(keep_filter_aux, keep))
+    for ext in ext_list:
         file_cp=basename+"."+ext
         if os.path.exists(file_cp):
             shutil.copy(file_cp, texfile_dir)
     os.chdir(texfile_dir)
 
 try:
+    # Send pid to ATP:
+    vim_remote_expr(servername, "atplib#PythonPID("+str(os.getpid())+")")
     cwd = getcwd()
     os.chdir(texfile_dir)
 
@@ -246,13 +283,6 @@ try:
 # this ensures that the aux, ... files are uptodate.
 
 # COPY FILES TO TEMP DIR
-    if not os.path.exists(os.path.join(texfile_dir, ".tmp")):
-            # This is the main tmp dir (./.tmp) 
-            # it will not be deleted by this script
-            # as another instance might be using it.
-            # it is removed by Vim on exit.
-        os.mkdir(os.path.join(texfile_dir,".tmp"))
-    tmpdir  = tempfile.mkdtemp(dir=os.path.join(texfile_dir,".tmp"),prefix="")
     debug_file.write("TMPDIR="+tmpdir+"\n")
     tmplog  = os.path.join(tmpdir,basename+".log")
     debug_file.write("TMPLOG="+tmplog+"\n")
@@ -281,19 +311,19 @@ try:
 
     output_exists=os.path.exists(os.path.join(texfile_dir,basename+output_ext))
     debug_file.write("OUTPUT_EXISTS="+str(output_exists)+":"+os.path.join(texfile_dir,basename+output_ext)+"\n")
-    latex=latex_progress_bar([cmd, '-interaction=nonstopmode', '-output-directory='+tmpdir]+tex_options_list+[texfile])
+    latex=latex_progress_bar([cmd, '-interaction=nonstopmode', '-output-directory='+tmpdir]+tex_options+[texfile])
     run  += 1
     latex.wait()
     vim_remote_expr(servername, "atplib#TexReturnCode('"+str(latex.returncode)+"')")
     os.chdir(tmpdir)
     if not output_exists:
-        copy_back_output()
+        copy_back_output(tmpdir)
         reload_xpdf()
 
 # AFTER FIRST TIME LOG FILE SHOULD EXISTS:
     if os.path.isfile(tmplog):
 
-        need_runs = 1
+        need_runs = [0]
 
         log_file  = open(tmplog, "r")
         log       = log_file.read()
@@ -308,20 +338,23 @@ try:
         for el in log_list:
             if el[0] != '' or el[1] != '':
                 citations       =True
-                if not biber:
-                    need_runs   =1
+                if biber:
+                    need_runs.append(1)
+                else:
+                    # Bibtex:
+                    need_runs.append(2)
             if el[2] != '':
                 labels          =True
             if el[3] != '':
                 makeidx         =True
-                need_runs       =2
+                need_runs.append(1)
 
         debug_file.write("citations="+str(citations)+"\n")
         debug_file.write("labels="+str(labels)+"\n")
         debug_file.write("makeidx="+str(makeidx)+"\n")
 
 # Scan for openout files to know if we are makeing: toc, lot, lof, thm
-        openout_list=re.findall("\\\\openout\d+\s*=\s*`([^']*)'",log)
+        openout_list=re.findall("\\\\openout\d+\s*=\s*`\"?([^'\"]*)\"?'",log)
         toc     =False
         lot     =False
         lof     =False
@@ -329,16 +362,17 @@ try:
         for el in openout_list:
             if re.search('\.toc$',el):
                 toc=True
-                need_runs=2
+                need_runs.append(1)
+                # We need only one more run, because of the 0-run.
             if re.search('\.lof$',el):
                 lof=True
-                need_runs=2
+                need_runs.append(1)
             if re.search('\.lot$',el):
                 lot=True
-                need_runs=2
+                need_runs.append(1)
             if re.search('\.thm$',el):
                 thm=True
-                need_runs=2
+                need_runs.append(1)
 
         debug_file.write("A0 need_runs="+str(need_runs)+"\n")
 
@@ -354,8 +388,11 @@ try:
         aux=aux_file.read()
         aux_file.close()
         bibtex=re.search('\\\\bibdata\s*{', aux)
+# This can be used to make it faster and use the old bbl file.
+# For this I have add a switch (bang).
+#         bibtex=re.search('No file '+basename+'\.bbl\.', log)
         if not bibtex:
-# Then search for biblatex package.
+# Then search for biblatex package. Alternatively, I can search for biblatex messages in log file.
             for line in open(texfile):
                 if re.match('[^%]*\\\\usepackage\s*(\[[^]]*\])?\s*{(\w\|,)*biblatex',line):
                     bibtex=True
@@ -382,6 +419,7 @@ try:
                     bibtex=subprocess.Popen([bibcmd, auxfile], stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
                     vim_remote_expr(servername, "atplib#BibtexPID('"+str(bibtex.pid)+"')")
                     vim_remote_expr(servername, "atplib#redrawstatus()")
+                    pids.append(bibtex.pid)
                     bibtex.wait()
                     vim_remote_expr(servername, "atplib#PIDsRunning(\"b:atp_BibtexPIDs\")")
                     bibtex_output=re.sub('"', '\\"', bibtex.stdout.read().decode())
@@ -394,6 +432,7 @@ try:
                     index=subprocess.Popen(['makeindex', idxfile], stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
                     vim_remote_expr(servername, "atplib#MakeindexPID('"+str(index.pid)+"')")
                     vim_remote_expr(servername, "atplib#redrawstatus()")
+                    pids.append(index.pid)
                     index.wait()
                     vim_remote_expr(servername, "atplib#PIDsRunning(\"b:atp_MakeindexPIDs\")")
                     index_returncode=index.returncode
@@ -401,12 +440,12 @@ try:
 
 # LATEX
             os.chdir(texfile_dir)
-            latex=latex_progress_bar([cmd, '-interaction=nonstopmode', '-output-directory='+tmpdir]+tex_options_list+[texfile])
+            latex=latex_progress_bar([cmd, '-interaction=nonstopmode', '-output-directory='+tmpdir]+tex_options+[texfile])
             run  += 1
             latex.wait()
             vim_remote_expr(servername, "atplib#CatchStatus('"+str(latex.returncode)+"')")
             reload_xpdf()
-            copy_back_output()
+            copy_back_output(tmpdir)
 
 #CONDITION
             log_file=open(tmplog, "r")
@@ -415,7 +454,7 @@ try:
 
 # Citations undefined|Label(s) may have changed
 #         log_list=re.findall('(C\n?i\n?t\n?a\n?t\n?i\n?o\n?n\n?s\s+u\n?n\n?d\n?e\n?f\n?i\n?n\n?e\n?d)|(L\n?a\n?b\n?e\n?l\(s\)\s+m\n?a\n?y\s+h\n?a\n?v\n?e\s+c\n?h\n?a\n?n\n?g\n?e\n?d)',log)
-            log_list=re.findall('(undefined references)|(Citations undefined)|(Label\(s\) may have changed)|(Writing index file)',log)
+            log_list=re.findall('(undefined references)|(Citations undefined)|(Label\(s\) may have changed)',log)
             citations       =False
             labels          =False
             for el in log_list:
@@ -429,7 +468,7 @@ try:
 
             debug_file.write(str(run)+"need_runs="+str(need_runs)+"\n")
 
-            condition = ( (citations and run <= need_runs) or labels or makeidx or run <= need_runs ) and run <= bound
+            condition = ( (citations and run <= max(need_runs)) or labels or makeidx or run <= max(need_runs) ) and run <= bound
             debug_file.write(str(run)+"condition="+str(condition)+"\n")
 
     # Start viewer: (reloading xpdf is done after each compelation) 
@@ -455,14 +494,14 @@ try:
         if start == 2:
             debug_file.write("SyncTex with "+str(viewer))
             vim_remote_expr(servername, "atplib#SyncTex()")
-    copy_back()
+    copy_back(tmpdir, latex.returncode)
 # else:
 # THERE IS NO LOG FILE AFTER FIRST TIME: exit with error.
 except Exception:
     error_str=re.sub("'", "''",re.sub('"', '\\"', traceback.format_exc()))
     traceback.print_exc(None, debug_file)
-    vim_remote_expr(servername, "atplib#Echo(\"[ATP:] error in makelatex.py, catched python exception:\n"+error_str+"[ATP info:] this error message is recorded in makelatex.py.log under g:atp_TempDir\",'echo','ErrorMsg')")
+    vim_remote_expr(servername, "atplib#Echo(\"[ATP:] error in makelatex.py, catched python exception:\n"+error_str+"[ATP info:] this error message is recorded in makelatex.log under g:atp_TempDir\",'echo','ErrorMsg')")
 
-os.chdir(cwd)
-debug_file.close()
-shutil.rmtree(tmpdir)
+debug_file.write("PIDS="+str(pids))
+vim_remote_expr(servername, "atplib#Echo('[ATP:] MakeLatex finished.', 'echomsg', 'Normal')")
+sys.exit(latex.returncode)
